+++ ReactAndroid/src/main/jni/react/fabric/Binding.cpp
@@ -468,12 +468,20 @@ void Binding::installFabricUIManager(
   auto toolbox = SchedulerToolbox{};
   toolbox.contextContainer = contextContainer;
   toolbox.componentRegistryFactory = componentsRegistry->buildRegistryFunction;
+
+  // TODO: (T130208323) runtimeExecutor should execute lambdas after
+  // main bundle eval, and bindingsInstallExecutor should execute before.
+  toolbox.bridgelessBindingsExecutor = std::nullopt;
   toolbox.runtimeExecutor = runtimeExecutor;
+
   toolbox.synchronousEventBeatFactory = synchronousBeatFactory;
   toolbox.asynchronousEventBeatFactory = asynchronousBeatFactory;
 
-  backgroundExecutor_ = JBackgroundExecutor::create("fabric_bg");
-  toolbox.backgroundExecutor = backgroundExecutor_;
+  if (reactNativeConfig_->getBool(
+          "react_fabric:enable_background_executor_android")) {
+    backgroundExecutor_ = JBackgroundExecutor::create("fabric_bg");
+    toolbox.backgroundExecutor = backgroundExecutor_;
+  }
 
   animationDriver_ = std::make_shared<LayoutAnimationDriver>(
       runtimeExecutor, contextContainer, this);