+++ packages/react-native-codegen/src/generators/components/GenerateEventEmitterH.js
@@ -285,9 +285,7 @@ module.exports = {
       .filter(Boolean)
       .reduce((acc, components) => Object.assign(acc, components), {});
 
-    const moduleComponentsWithEvents = Object.keys(moduleComponents).filter(
-      componentName => moduleComponents[componentName].events.length > 0,
-    );
+    const moduleComponentsWithEvents = Object.keys(moduleComponents);
 
     const fileName = 'EventEmitters.h';