+++ packages/react-native-gradle-plugin/src/main/kotlin/com/facebook/react/ReactPlugin.kt
@@ -46,7 +46,8 @@ class ReactPlugin : Plugin<Project> {
       
       ********************************************************************************
       
-      """.trimIndent())
+      """
+              .trimIndent())
       exitProcess(1)
     }
   }